Added COLLATE SQL_Latin1_General_CP1_CS_AS to tag and version of tags table

+++ Analyst_sln/Analyst.DBAccess/recreation_script.sql
@@ -1,6 +1,11 @@
 
+DECLARE @dbname nvarchar(128)
+SET @dbname = N'Analyr'
+
 use [master];
-drop database Analyst
+IF (EXISTS (SELECT name FROM master.dbo.sysdatabases WHERE ('[' + name + ']' = @dbname OR name = @dbname)))
+	PRINT 'Droping Analyst'
+	drop database Analyst
 go
 
 use [master];
@@ -38,7 +43,12 @@ go
 use [Analyst];
 EXEC sp_addrolemember N'db_owner', N'analyst_usr'
 
-
+----------------------------------------------------------------------
+/*
+--no funciona porque tiene el indice
+use [Analyst];
+ALTER TABLE [dbo].[EdgarDatasetTags] ALTER COLUMN [Tag] VARCHAR(256) COLLATE SQL_Latin1_General_CP1_CS_AS NOT NULL
+*/
 ----------------------------------------------------------------------
 
 /*
